requirements.txt
pyspellchecker
regex